✅ Valid Perfect Square | Easy | Java | Runtime: 0 ms (100.00%) | Memory: 41.9 MB | 2026-08-01

+++ leetcode/easy/0367-valid-perfect-square/solution.java
@@ -5,7 +5,7 @@ class Solution {
        
        while(low<=high){
         int mid=low+(high-low)/2;
-        long sq=mid*mid;
+        long sq=(long)mid*mid;
         if(sq==num) return true;
         else if(sq>num) high=mid-1;
         else low=mid+1;
